homework_06/pyproject.toml
[project]
name = "homework_06"
version = "0.1.0"
description = ""
authors = [
    {name = "PristenCityMan",email = "166030137+PristenCityMan@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

